City

+++ MStore.Application/Dtos/CatalogDtos/ProductWarehouseInventoryDto/GetProductWarehouseInventoryDto.cs
@@ -16,11 +16,11 @@ namespace MStore.Application.Dtos.CatalogDtos.ProductWarehouseInventoryDto
         /// <summary>
         /// Gets or sets the stock quantity
         /// </summary>
-        public int StockQuantity { get; set; }
+        public Decimal StockQuantity { get; set; }
 
         /// <summary>
         /// Gets or sets the reserved quantity (ordered but not shipped yet)
         /// </summary>
-        public int ReservedQuantity { get; set; }
+        public Decimal ReservedQuantity { get; set; }
     }
 }
